feat: add flex layout to sources list. Refactor type defintion code location and imports
